RUBY-2213 rename server_discovery_and_monitoring to sdam for consistency
